% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/backTestVaR.R
\name{backtestVaR.GARCH}
\alias{backtestVaR.GARCH}
\title{GARCH Model VaR Backtest}
\usage{
backtestVaR.GARCH(garch, p = c(0.95, 0.99), nAhead = 1, refitEvery = 25,
  window = 100)
}
\arguments{
\item{garch}{uvGARCH object create via \code{\link{uvGARCH}}}

\item{p}{confidence level for the VaR estimate.}

\item{nAhead}{number of steps ahead to forecast. (nAhead = 1 only supported)}

\item{refitEvery}{number of periods the mode is refit}

\item{window}{size of the moving window in the rolling VaR estimate.}
}
\description{
Function for rolling estimate of GARCH model and VaR backtest
}
\author{
Ross Bennett
}
\seealso{
\code{\link[rugarch]{ugarchroll}}
}

