% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/backTestVaR.R
\name{backtestVaR}
\alias{backtestVaR}
\title{Backtest Value-at-Risk (VaR)}
\usage{
backtestVaR(R, window = 100, p = 0.95, method = "historical",
  bootstrap = FALSE, replications = 1000, bootParallel = FALSE)
}
\arguments{
\item{R}{xts or zoo object of asset returns}

\item{window}{size of the moving window in the rolling VaR estimate.}

\item{p}{confidence level for the VaR estimate.}

\item{method}{method for the VaR calculation. Valid choices are "modified", "guassian", "historical", and "kernel"}

\item{bootstrap}{TRUE/FALSE use the bootstrap estimate for the VaR calculation, (default FALSE).}

\item{replications}{number of bootstrap replications.}

\item{bootParallel}{TRUE/FALSE run the bootstrap in parallel, (default FALSE).}
}
\description{
Backtesting Value-at-Risk estimate over a moving window.
}
\details{
The size of the moving window is set with the \code{window} argument. For
example, if the window size is 100, periods 1:100 are used to estimate the
VaR level for period 101.
}
\examples{
data(crsp_weekly)
R <- largecap_weekly[, 1]
backtest <- backtestVaR(R, window=100, p=0.95, method=c("gaussian", "historical", "modified"))
backtest

head(getVaREstimates(backtest))
head(getVaRViolations(backtest))
}
\author{
Ross Bennett
}
\seealso{
\code{\link[PerformanceAnalytics]{VaR}}, \code{\link{bootVaR}}
}

