% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/riskMetricsAndHedges.R
\name{bondConvexity}
\alias{bondConvexity}
\title{Calculate the convexity of a fixed rate coupon bond}
\usage{
bondConvexity(bond, discountCurve)
}
\arguments{
\item{bond}{a \code{bond} object in discountFactorArbitrage}

\item{discountCurve}{vector of discount rates}
}
\value{
convexity of the bond
}
\description{
This function estimates the convexity of a fixed rate coupon bond
given the discount curve and bond data.
}
\examples{
time = seq(from=0.5, to=2, by=0.5)
DF = rbind(0.968,0.9407242,0.9031545,0.8739803)
bond = bondSpec(time, face=100, m=2, couponRate = 0.0475)
convexity = bondConvexity(bond,DF)
}
\author{
Thomas Fillebeen
}

