% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/riskMetricsAndHedges.R
\name{bondDuration}
\alias{bondDuration}
\title{Calculate the duration of a bond}
\usage{
bondDuration(bond, discountCurve, percentChangeYield = 0,
  type = c("modified", "macaulay"))
}
\arguments{
\item{bond}{a \code{bond} object created with \code{\link{bondSpec}}}

\item{discountCurve}{vector of discount rates}

\item{percentChangeYield}{optional elasticity measure}

\item{type}{specify modified or macaulay duration}
}
\value{
duration of the bond
}
\description{
Estimate the macaulay or modified duration of a fixed rate coupon bond
given the discount curve and bond data. The duration is calculated
using the continuously compounded yield
}
\examples{
time = seq(from=0.5, to=2, by=0.5)
DF = rbind(0.968,0.9407242,0.9031545,0.8739803)
bond = bondSpec(time, face=100, m=2, couponRate = 0.0475)
mcDuration = bondDuration(bond,DF, type="macaulay")
modDuration = bondDuration(bond,DF, type="modified")
}
\author{
Thomas Fillebeen and Jaiganesh Prabhakaran
}

