% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/discountFactorArbitrage.R
\name{bondFullPrice}
\alias{bondFullPrice}
\title{bondFullPrice}
\usage{
bondFullPrice(bond, yield, cashFlowPd, t0, t1, currentDate)
}
\arguments{
\item{bond}{is a bondSpec object created by \code{\link{bondSpec}}}

\item{yield}{is the yield on the bond}

\item{cashFlowPd}{cash flow period}

\item{t0}{previous coupon date}

\item{t1}{next coupon period}

\item{currentDate}{current date}
}
\value{
price of the bond: clean, dirty and accrued interest
}
\description{
Estimate price of bond w/ acrrued interest
The present value of a bond's cash flows should be equated or
compared with its full price, with the amount a purchaser actually
pays to purchase those cash flows. The flat price is denoted by p, accrued
interest is AI, the present value of the cash flows by PV, and the
full price by P:
P=p+AI=PV
}
\details{
This function calculates the price of a fixed rate coupon bond given coupon rate, yield,
compoundPd, cashFlowPd, face value, previous coupon date, next coupon date.
}
\examples{
t0 = as.Date("2013-08-15")
t1 = as.Date("2014-02-15")
tn = as.Date("2013-10-04")
currentDate = tn
bond = bondSpec(face=100, m=2, couponRate = 0.0475)
y1 = 0.00961
bondFullPrice(bond, y1, 8, t0, t1, tn)$clean
bondFullPrice(bond, y1, 8, t0, t1, tn)$dirty
bondFullPrice(bond, y1, 8, t0, t1, tn)$accruedInterest
}
\author{
Thomas Fillebeen
}

