% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/discountFactorArbitrage.R
\name{bondPrice}
\alias{bondPrice}
\title{Estimate price of bond}
\usage{
bondPrice(bond, discountCurve)
}
\arguments{
\item{bond}{a \code{bondSpec} object}

\item{discountCurve}{vector of discount rates}
}
\value{
price of the bond
}
\description{
This function calculates the price of a fixed rate coupon bond given the
discount curve and bond data. First it converts the discountCurve into CF
}
\examples{
time = seq(from=0.5, to=2, by=0.5)
bond = bondSpec(time, face=100, m=2, couponRate = 0.0475)
DF = rbind(0.968,0.9407242,0.9031545,0.8739803)
price = bondPrice(bond,DF)
}
\author{
Thomas Fillebeen
}

