% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/riskMetricsAndHedges.R
\name{bondYTM}
\alias{bondYTM}
\title{Calculate the yield to maturity of a bond}
\usage{
bondYTM(bond, discountCurve)
}
\arguments{
\item{bond}{a \code{bond} object}

\item{discountCurve}{vector of discount rates}
}
\value{
yield to maturity of the bond
}
\description{
This function calculates the yield to maturity of a fixed rate coupon bond
given the discount curve and bond data.
}
\examples{
time = seq(from=0.5, to=2, by=0.5)
DF = rbind(0.968,0.9407242,0.9031545,0.8739803)
bond = bondSpec(time, face=100, m=2, couponRate = 0.0475)
bondYTM(bond,DF)
}
\author{
Thomas Fillebeen
}

