% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/boot.R
\name{bootCor}
\alias{bootCor}
\title{Bootstrap Correlation}
\usage{
bootCor(R, ..., replications = 1000)
}
\arguments{
\item{R}{xts object or matrix of asset returns}

\item{replications}{number of bootstrap replications.}

\item{\dots}{passthrough parameters to \code{\link[stats]{cor}}}
}
\description{
Bootstrap the correlation of an xts object or matrix of asset returns
}
\examples{
data(crsp_weekly)
R <- largecap_weekly[,1:4]
bootCor(R[,1:2])
bootCor(R[,1:2], method="kendall")
bootCor(R)
}
\author{
Ross Bennett
}
\seealso{
\code{\link[stats]{cor}}
}

