% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/boot.R
\name{bootFUN}
\alias{bootFUN}
\title{Bootstrap}
\usage{
bootFUN(R, FUN = "mean", ..., replications = 1000)
}
\arguments{
\item{R}{xts object or matrix of data passed to \code{FUN}.}

\item{FUN}{the function to be applied.}

\item{replications}{number of bootstrap replications.}

\item{\dots}{optional arguments to \code{FUN}.}
}
\description{
Bootstrap a function
}
\details{
\code{R} is the data passed to \code{FUN}. \code{FUN} must have \code{x} or
\code{R} as arguments for the data. For example, see the functions linked to
in the 'See Also' section. Care must be taken when using \code{bootFUN} on
multivariate data. This function is designed to only accept univariate
(i.e. ncol(R) = 1) data, however is made to work with bivariate data for
\code{bootCor} and \code{bootCov}. For multivariate data, a wrapper function
should be written to apply the bootstrap function to each column of data.
}
\author{
Ross Bennett
}
\seealso{
\code{\link{bootMean}}, \code{\link{bootSD}}, \code{\link{bootStdDev}},
\code{\link{bootSimpleVolatility}}, \code{\link{bootCor}}, \code{\link{bootCov}},
\code{\link{bootVaR}}, \code{\link{bootES}}
}

