% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/boot.R
\name{bootMean}
\alias{bootMean}
\title{Bootstrap Mean}
\usage{
bootMean(R, ..., replications = 1000)
}
\arguments{
\item{R}{xts object or matrix of asset returns}

\item{replications}{number of bootstrap replications.}

\item{\dots}{passthrough parameters to \code{\link[base]{mean}}}
}
\description{
Bootstrap the mean of an xts object or matrix of asset returns
}
\examples{
data(crsp_weekly)
R <- largecap_weekly[,1:4]
bootMean(R[,1])
bootMean(R)
}
\author{
Ross Bennett
}
\seealso{
\code{\link[base]{mean}}
}

