% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/discountFactorArbitrage.R
\name{compoundingRate}
\alias{compoundingRate}
\title{Estimate continuously conpounding rate to be used in term structure}
\usage{
compoundingRate(dat, initialDate = as.Date("1995-05-15"), m, face = 100)
}
\arguments{
\item{dat}{is a dataset with cusip, issueDate, MaturityDate, Name, Coupon, Bid/Ask}

\item{initialDate}{is the date when the estimation should be conducted: date of reference}

\item{m}{compounding frequency}

\item{face}{face value}
}
\value{
continuously compounding rates
}
\description{
This function calculates the continuously compounding rate given an initial dataset
with specific format, date of reference coumpounding frequency, and face value
}
\author{
Thomas Fillebeen
}

