\name{consumption}
\docType{data}
\alias{consumption}
\title{consumption data}
\description{
Monthly consumption data of what?
}
\usage{consumption}
\details{
Where did we get the consumption data?
What are the details of the consumption data? Consumption of what?

see \url{TODO}
}
\format{xts object with monthly observations}
\source{\url{TODO}}
\references{

}
\examples{
data(consumption)
head(consumption)
}
\keyword{datasets}
