\name{countViolations}
\alias{countViolations}
\title{Count backtesting VaR}
\usage{
  countViolations(object, temp, initialWindow, CI)
}
\arguments{
  \item{backTestVaR}{object created by
  \code{\link{backTestVaR}}}

  \item{initialWindow}{}

  \item{CI}{}

  \item{temp}{}
}
\description{
  Description of countBacktesting VaR
}

