% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/discountFactorArbitrage.R
\name{discountFactor}
\alias{discountFactor}
\title{Estimate discountFactor}
\usage{
discountFactor(price, cashFlow)
}
\arguments{
\item{price}{of a bond}

\item{cashFlow}{of a bond}
}
\value{
discount factors
}
\description{
This function calculates the discountFactor (DF) given price
and cashFlows.
}
\examples{
cashFlow = rbind(c(100+(1+1/4)/2,0,0),c((4 +7/8)/2,100+(4+7/8)/2,0),c((4+1/2)/2,(4+1/2)/2,100+(4+1/2)/2))
# Created Price of the bond
price = matrix(c(100.550, 104.513, 105.856), ncol=1)
DF = discountFactor(price, cashFlow)
}
\author{
Thomas Fillebeen
}

