% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/efficient_frontier.R
\name{efficientFrontier}
\alias{efficientFrontier}
\title{Efficient Frontier}
\usage{
efficientFrontier(R, nPortfolios = 25, minBox = 0, maxBox = 1,
  groupList = NULL, groupMin = NULL, groupMax = NULL)
}
\arguments{
\item{R}{xts object of asset returns}

\item{nPortfolios}{number of portfolios to generate along efficient frontier}

\item{minBox}{box constraint minimum}

\item{maxBox}{box constraint maximum}

\item{groupList}{list specifying asset groups}

\item{groupMin}{group constraint minimum}

\item{groupMax}{group constraint maximum}
}
\description{
Generate portfolios along an efficient frontier.
}
\details{
This is a wrapper function for code in PortfolioAnalytics to initialize a
portfolio and create an efficint frontier in mean - standard deviation space.
Box constraints and group constraints are supported for constrained
optimization to generate portfolios along the efficient frontier.
}

