% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/efficient_frontier.R
\name{efficientFrontierTwoAsset}
\alias{efficientFrontierTwoAsset}
\title{Efficient Frontier for Portfolio of Two Assets}
\usage{
efficientFrontierTwoAsset(R1, R2, sigma1, sigma2, rho, nPortfolios = 25,
  rf = 0, allowShorting = FALSE, weights = NULL)
}
\arguments{
\item{R1}{expected return for asset 1}

\item{R2}{expected return for asset 2}

\item{sigma1}{standard deviation for asset 1}

\item{sigma2}{standard deviation for asset 2}

\item{rho}{correlation coefficient between asset 1 and asset 2}

\item{nPortfolios}{number of portfolios to generate along efficient frontier}

\item{rf}{risk free rate}

\item{allowShorting}{TRUE/FALSE to allow short sales}

\item{weights}{vector of weights for fraction of asset 1 used to generate
the efficient frontier. If the \code{weights} argument is specified, the
\code{nPortfolios} and \code{allowShorting} arguments are ignored.}
}
\description{
Create an efficient frontier for a portfolio of two risky assets
}
\details{
This is a specialized function to generate points along the efficient
frontier for a portfolio of two assets following the equations presented in
Chapter 3: Delineating Efficient Portfolios.
}

