% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/monte_carlo.R
\name{endingPrices}
\alias{endingPrices}
\title{Ending Prices of Monte Carlo Simulation}
\usage{
endingPrices(mc)
}
\arguments{
\item{mc}{monte carlo object created with \code{monteCarlo}}
}
\value{
vector ending prices
}
\description{
Get the ending prices, i.e. terminal values, of a monte carlo simulation
}
\examples{
library(GARPFRM)

mc <- monteCarlo(0.05, 0.25, 500, 1, 52, 10)
ep <- endingPrices(mc)
}

