\name{fama_french_factors}
\docType{data}
\alias{fama_french_factors}
\title{Factors for the Fama/French 3 factor model}
\description{
Weekly observations of Fama/French factors.
}
\usage{fama_french_factors}
\details{
see \url{http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/Data_Library/f-f_factors.html}
}
\format{xts object with weekly observations}
\source{\url{http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html}}
\references{

}
\examples{
data(fama_french_factors)
head(fama_french_factors)
}
\keyword{datasets}


