\name{fcstGarch11.DCCfit}
\alias{fcstGarch11.DCCfit}
\title{#Forecast GARCH(1,1)
#
#Description of forecast GARCH(1,1)
#
#}
\usage{
\method{fcstGarch11}{Dccfit #}(object, window = 100)
}
\arguments{
  \item{garch11}{object created by \code{\link{GARCH(1,1)}}
  #}

  \item{window}{is the forecast window (default is set to
  window = 100) #}

  \item{object}{garch11}

  \item{window}{}
}
\description{
#Forecast GARCH(1,1) # #Description of forecast GARCH(1,1)
# #
}

