\name{fcstGarch11}
\alias{fcstGarch11}
\title{Forecast GARCH(1,1)}
\usage{
fcstGarch11(object, window)
}
\arguments{
  \item{garch11}{object created by
  \code{\link{GARCH(1,1)}}}

  \item{window}{is the forecast window (default is set to
  window = 100)}
}
\description{
Description of forecast GARCH(1,1)
}

