\name{garch11}
\alias{garch11}
\title{GARCH(1,1)}
\usage{
garch11(R, model = "sGARCH", distribution.model = "norm")
}
\arguments{
  \item{R}{GARCH(1,1)}

  \item{model}{“sGARCH”, “fGARCH”, “eGARCH”, “gjrGARCH”,
  “apARCH” and “iGARCH” and “csGARCH”}

  \item{distribution.model.}{Valid choices are “norm” for
  the normal distibution, “snorm” for the skew-normal
  distribution, “std” for the student-t, “sstd” for the
  skew-student, “ged” for the generalized error
  distribution, “sged” for the skew-generalized error
  distribution, “nig” for the normal inverse gaussian
  distribution, “ghyp” for the Generalized Hyperbolic, and
  “jsu” for Johnson's SU distribution.}
}
\description{
Description of GARCH(1,1)
}

