% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/capm.R
\name{getAlphas}
\alias{getAlphas}
\title{CAPM alphas}
\usage{
getAlphas(object)
}
\arguments{
\item{object}{a capm object created by \code{\link{CAPM}}}
}
\description{
Extract the computed alphas (intercept) from the fitted CAPM object.
}
\author{
Thomas Fillebeen
}

