% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/capm.R
\name{getBetas}
\alias{getBetas}
\title{CAPM betas}
\usage{
getBetas(object)
}
\arguments{
\item{object}{a capm object created by \code{\link{CAPM}}}
}
\description{
Extract the computed alpha (intercept) from the CAPM object.
}
\author{
Thomas Fillebeen
}

