% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/EWMA.R
\name{getCor}
\alias{getCor}
\title{EWMA Correlation}
\usage{
getCor(EWMA, assets)
}
\arguments{
\item{EWMA}{an EWMA object created by \code{\link{EWMA}}}

\item{assets}{character vector or numeric vector. The assets can be
specified by name or index.}
}
\description{
Extract the correlation of two assets from an \code{mvEWMAcor} object
}
\examples{
data(crsp_weekly)
mvR <- largecap_weekly[,1:4]
lambda <- 0.94
initialWindow <- 150
cor_mv <- EWMA(mvR, lambda, initialWindow, type="correlation")
# Extract the estimated correlation between ORCL and MSFT
tail(getCor(cor_mv, assets=c("ORCL", "MSFT")))
}
\author{
Ross Bennett and Thomas Fillebeen
}

