% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/EWMA.R
\name{getEstimate}
\alias{getEstimate}
\title{Get Estimated Values}
\usage{
getEstimate(model, ...)
}
\arguments{
\item{model}{fitted model (currently only EWMA)}

\item{\dots}{passthrough parameters (not currently used)}
}
\value{
model estimate
}
\description{
Get the estimated values from the model
}
\author{
Ross Bennett
}

