% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/riskMetricsAndHedges.R
\name{getLoadings}
\alias{getLoadings}
\title{Retrieve PCA loadings}
\usage{
getLoadings(object)
}
\arguments{
\item{object}{is a pca object created by \code{\link{PCA}}}
}
\description{
Retrieve PCA loadings
}
\author{
Thomas Fillebeen
}

