% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/capm.R
\name{getStatistics}
\alias{getStatistics}
\title{CAPM statistics}
\usage{
getStatistics(object)
}
\arguments{
\item{object}{a capm object created by \code{\link{CAPM}}.}
}
\description{
Extract the standard error, t-values, and p-values from the CAPM object.
}
\details{
The t-statistic and corresponding two-sided p-value are calculated differently
for the alpha and beta coefficients.
\itemize{
  \item{alpha}{ the t-statistic and corresponding p-value are calculated to
  test if alpha is significantly different from 0.
  \deqn{
    H0: \alpha = 0
  }
  }
  \item{beta}{ the t-statistic and corresponding p-value are calculated to
  test if beta is significantly different from 1.
  \deqn{
    H0: \beta = 1
  }
  }
}
}
\author{
Thomas Fillebeen
}

