\name{getVar}
\alias{getVar}
\title{EWMA Variance}
\usage{
  getVar(EWMA, assets)
}
\arguments{
  \item{object}{an EWMA object created by \code{EWMA}}

  \item{assets}{character vector or numeric vector. The
  assets can be specified by name or index.}
}
\description{
  Extract the Variance of an asset from an \code{EWMA}
  object
}

