\name{getVaREstimates}
\alias{getVaREstimates}
\title{VaR Estimates
Extract VaR Estimates from a VaR Backtest}
\usage{
  getVaREstimates(object, ...)
}
\arguments{
  \item{object}{an object created by
  \code{\link{backtestVaR}}.}

  \item{\dots}{not currently used}
}
\value{
  xts object of unconditional VaR estimates
}
\description{
  VaR Estimates Extract VaR Estimates from a VaR Backtest
}
\author{
  Ross Bennett
}
\seealso{
  \code{\link{backtestVaR}}
}

