% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/riskMetricsAndHedges.R
\name{getWeights}
\alias{getWeights}
\title{Retrieve PCA weights}
\usage{
getWeights(object)
}
\arguments{
\item{object}{is a pca object created by \code{\link{PCA}}}
}
\description{
Retrieve PCA weights
}
\author{
Thomas Fillebeen
}

