% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/capm.R
\name{hypTest}
\alias{hypTest}
\title{CAPM Hypothesis Test}
\usage{
hypTest(object, significanceLevel)
}
\arguments{
\item{object}{a capm object created by \code{\link{CAPM}}}

\item{significanceLevel}{confidence level}
}
\value{
TRUE if the null hypothesis is rejected (i.e. the estimated coefficient is significant)
FALSE if the null hypothesis cannot be rejected (i.e. the estimated coefficient is not significant)
}
\description{
Test the CAPM coefficients for significance.
}
\details{
This function tests the significance of the coefficients (alpha and beta)
estimated by the CAPM.

#' The t-statistic and corresponding two-sided p-value are calculated differently
for the alpha and beta coefficients
\itemize{
  \item{alpha}{ the t-statistic and corresponding p-value are calculated to
  test if alpha is significantly different from 0.
  \deqn{
    H0: \alpha = 0
  }
  }
  \item{beta}{ the t-statistic and corresponding p-value are calculated to
  test if beta is significantly different from 1.
  \deqn{
    H0: \beta = 1
  }
  }
}

If the p-value is less than the coefficients confidence level, the null
hypothesis is rejected meaning that the coefficients is significant. If
the p-value is greater than the specified confidence level, the null
hypothesis cannot be rejected.
}
\author{
Thomas Fillebeen
}
\seealso{
\code{\link{getStatistics}}
}

