% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/options.R
\name{impliedVolatility}
\alias{impliedVolatility}
\title{Implied Volatility}
\usage{
impliedVolatility(option, price, lower = 0, upper = 0.5, ...)
}
\arguments{
\item{option}{an \code{option} object created with \code{\link{optionSpec}}}

\item{price}{market price of the option}

\item{lower}{the lower bound of implied volatility to search}

\item{upper}{the upper bound of implied volatility to search}

\item{\dots}{any passthrough parameters to \code{\link{impliedVolBS}}}
}
\value{
implied volatility estimate
}
\description{
Compute the implied volatility of a european option using the
Black-Scholes-Merton model.
}
\details{
A bisection algorithm is used to compute the implied volatility
of a European option priced with the Black-Scholes-Merton model
}
\author{
Ross Bennett
}

