% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/options.R
\name{impliedVolBS}
\alias{impliedVolBS}
\title{Implied Volatility Bisection Method}
\usage{
impliedVolBS(vol_range, S0, K, r, q, ttm, P_mkt, type,
  tol = .Machine$double.eps, max_it = 200)
}
\arguments{
\item{vol_range}{c(lower, upper) the lower and upper bounds of the implied
volatility range to search}

\item{S0}{underlying asset price}

\item{K}{strike price}

\item{r}{risk free rate}

\item{q}{continuous dividend yield rate for options on stocks or stock
indices paying a dividend. Also the foreign risk free rate for options on
currencies}

\item{ttm}{time to maturity, the life of the option, measured in years}

\item{P_mkt}{market price}

\item{type}{type of the option; "call" or "put"}

\item{tol}{tolerance used for stopping criteria}

\item{max_it}{maximum number of iterations}
}
\value{
implied volatility estimate
}
\description{
Bisection method to compute the implied volatility of a european option
using the Black-Scholes-Merton model.
}
\details{
A bisection algorithm is used to compute the implied volatility
of a European option priced with the Black-Scholes-Merton model
}
\author{
Ross Bennett
}

