\name{microcap.ts}
\docType{data}
\alias{microcap.ts}
\title{CRSP Micro Cap Monthly Returns}
\description{
Monthly returns data of a selection of micro cap stocks.
}
\usage{microcap.ts}
\details{
CRSP Data used in GARPFRM and related publications with the kind permission of the Center for Research in Security Prices.

see \url{http://www.crsp.com/about-crsp}
}
\format{xts object with monthly observations}
\source{\url{http://www.crsp.com}}
\references{

}
\examples{
data(crsp.short)
head(microcap.ts)
}
\keyword{datasets}


