\name{microcap_weekly}
\docType{data}
\alias{microcap_weekly}
\title{CRSP Micro Cap Weekly Returns}
\description{
Weekly returns data of a selection of large cap stocks.
}
\usage{microcap_weekly}
\details{
CRSP Data used in GARPFRM and related publications with the kind permission of the Center for Research in Security Prices.

see \url{http://www.crsp.com/about-crsp}
}
\format{xts object with weekly observations}
\source{\url{http://www.crsp.com}}
\references{

}
\examples{
data(crsp_weekly)
head(microcap_weekly)
}
\keyword{datasets}


