% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Modified.bondDuration}
\alias{Modified.bondDuration}
\title{Calculate the modified duration of a bond}
\usage{
Modified.bondDuration(bond, discountCurve, percentChangeYield = 0)
}
\arguments{
\item{bond}{a \code{bond} object in discountFactorArbitrage}

\item{discountCurve}{vector of discount rates}

\item{percentChangeYield}{optional elasticity measure}
}
\value{
Modified Duration of the bond
}
\description{
The function estimates modified duration of a fixed rate coupon bond
given the discount curve and bond data. The modified duration is calculated
using the continuously compounded yield
}
\examples{
time = seq(from=0.5, to=2, by=0.5)
DF = rbind(0.968,0.9407242,0.9031545,0.8739803)
bond = bondSpec(time, face=100, m=2, couponRate = 0.0475)
mDuration = Modified.bondDuration(bond,DF)
}
\author{
Jaiganesh Prabhakaran
}

