% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/options.R
\name{optionSpec}
\alias{optionSpec}
\title{Option Specification}
\usage{
optionSpec(style = c("european", "american"), type = c("call", "put"),
  S0 = 100, K = 100, maturity = 1, r = 0.05, volatility = 0.2,
  q = 0)
}
\arguments{
\item{style}{style of the option, e.g. european, american, etc.}

\item{type}{type of the option; "call" or "put"}

\item{S0}{underlying asset price}

\item{K}{strike price}

\item{maturity}{the life of the option, measured in years}

\item{r}{risk free rate}

\item{volatility}{volatility of the underlying asset price}

\item{q}{continuous dividend yield rate for options on stocks or stock
indices paying a dividend. Also the foreign risk free rate for options on
currencies}
}
\value{
an object of class "option" with the parameters that specify the option
}
\description{
Specify parameters of an option
}
\examples{
am.call <- optionSpec(style="american", type="call")
euro.call <- optionSpec(style="european", type="call", S0=30, K=30, maturity=1, r=0.05, volatility=0.25, q=0)
}
\author{
Ross Bennett
}

