% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/backTestVaR.R
\name{plot.backtestVaR}
\alias{plot.backtestVaR}
\title{Plotting for VaR Backtest}
\usage{
\method{plot}{backtestVaR}(x, y, ..., pch = NULL, main = "VaR Backtest",
  ylim = NULL, colorset = NULL, legendLoc = NULL, legendCex = 0.8)
}
\arguments{
\item{x}{backtestVaR object created with \code{\link{backtestVaR}}.}

\item{y}{not used.}

\item{pch}{plotting 'character' for the violation points, same as in
\code{\link{plot}}. If NULL, violation points will not be plotted.}

\item{main}{the main title}

\item{ylim}{limits for the y-axis, same as in \code{\link{plot}}.}

\item{colorset}{colorset for plotting the VaR forecasts. The length of
colorset should be equal to the number of VaR methods.}

\item{legendLoc}{legend location. If NULL, no legend is plotted.}

\item{legendCex}{numerical value giving the amount by which the legend.
text should be magnified relative to the default.}

\item{\dots}{passthrough parameters to \code{\link{plot}}.}
}
\description{
Plotting method for VaR Backtest
}
\examples{
data(crsp_weekly)
R <- largecap_weekly[, 1]
backtest <- backtestVaR(R, window=100, p=0.95, method=c("gaussian", "historical", "modified"))
plot(backtest, pch=18, legendLoc="topright")
}

