\name{plot.capm_mlm}
\alias{plot.capm_mlm}
\title{Plotting method for CAPM}
\usage{
  \method{plot}{capm_mlm} (x, y, ..., main = "CAPM")
}
\arguments{
  \item{x}{a capm object created by \code{\link{CAPM}}.}

  \item{y}{not used}

  \item{\dots}{passthrough parameters to
  \code{\link{plot}}.}

  \item{main}{a main title for the plot}
}
\description{
  Plot a fitted CAPM object
}
\author{
  Thomas Fillebeen
}

