\name{plot.capm_uv}
\alias{plot.capm_uv}
\title{Plotting method for CAPM}
\usage{
\method{plot}{capm_uv}(x, y, ..., main = "CAPM")
}
\arguments{
  \item{x}{a capm object created by \code{\link{CAPM}}.}

  \item{y}{not used}

  \item{\dots}{passthrough parameters to
  \code{\link{plot}}.}

  \item{main}{a main title for the plot}
}
\description{
Plot a fitted CAPM object
}
\author{
Thomas Fillebeen
}

