% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/efficient_frontier.R
\name{portReturnTwoAsset}
\alias{portReturnTwoAsset}
\title{Portfolio Return for a Portfolio of Two Assets}
\usage{
portReturnTwoAsset(R1, R2, X1)
}
\arguments{
\item{R1}{expected return for asset 1}

\item{R2}{expected return for asset 2}

\item{X1}{fraction of portfolio invested in asset 1}
}
\value{
portfolio expected return
}
\description{
Calculate the portfolio return for a portfolio of two assets
}
\details{
This is a specialized function to calculate the return of a portfolio of two
assets following the equations presented in
Chapter 3: Delineating Efficient Portfolios.
}

