% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/efficient_frontier.R
\name{portSDTwoAsset}
\alias{portSDTwoAsset}
\title{Portfolio Standard Deviation for a Portfolio of Two Assets}
\usage{
portSDTwoAsset(R1, R2, X1, sigma1, sigma2, rho)
}
\arguments{
\item{R1}{expected return for asset 1}

\item{R2}{expected return for asset 2}

\item{X1}{fraction of portfolio invested in asset 1}

\item{sigma1}{standard deviation for asset 1}

\item{sigma2}{standard deviation for asset 2}

\item{rho}{correlation coefficient between asset 1 and asset 2}
}
\value{
portfolio expected return
}
\description{
Calculate the standard deviation for a portfolio of two assets
}
\details{
This is a specialized function to calculate the standard deviation of a
portfolio of two assets following the equations presented in
Chapter 3: Delineating Efficient Portfolios.
}

