\name{prices}
\docType{data}
\alias{prices}
\title{Equity Prices}
\description{
Weekly observations of adjusted close prices for SPY, AAPL, XOM, GOOG, MSFT, and GE.
}
\usage{prices}
\details{
Data obtained via yahoo finance api.
see \url{http://finance.yahoo.com}
}
\format{xts object with weekly observations}
\source{\url{http://finance.yahoo.com}}
\references{

}
\examples{
data(prices)
head(prices)
}
\keyword{datasets}


