% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/EWMA.R
\name{realizedCov}
\alias{realizedCov}
\title{Realized Covariance}
\usage{
realizedCov(R, n)
}
\arguments{
\item{R}{xts object of asset returns}

\item{n}{number of periods used to calculate realized volatility}
}
\value{
xts object of realized covariance
}
\description{
Calculate realized covariance
}
\details{
Realized covariance is calculated using the previous \code{n} periods.
}
\examples{
data(crsp_weekly)
R <- largecap_weekly[, 1:2]
# Calculate realized covariance
realizedCovariance <- realizedCov(R, 10)
head(realizedCovariance)
}
\author{
Ross Bennett
}

