% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/EWMA.R
\name{realizedVol}
\alias{realizedVol}
\title{Realized Volatility}
\usage{
realizedVol(R, n)
}
\arguments{
\item{R}{xts object of asset returns}

\item{n}{number of periods used to calculate realized volatility}
}
\value{
xts object of realized volatility
}
\description{
Calculate realized volatility
}
\details{
Realized volatility is defined as the standard deviation of using the
previous \code{n} periods.
}
\examples{
data(crsp_weekly)
R <- largecap_weekly[, 1]
# Calculate realized volatility
realizedVolatility <- realizedVol(R[,1], 10)
head(realizedVolatility)
}
\author{
Ross Bennett
}

