\name{returns}
\docType{data}
\alias{returns}
\title{Equity Returns}
\description{
Weekly observations of total returns for SPY, AAPL, XOM, GOOG, MSFT, and GE.
}
\usage{prices}
\details{
Data obtained via yahoo finance api.
see \url{http://finance.yahoo.com}
}
\format{xts object with weekly observations}
\source{\url{http://finance.yahoo.com}}
\references{

}
\examples{
data(returns)
head(returns)
}
\keyword{datasets}


