% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rollFUN.R
\name{rollCov}
\alias{rollCov}
\title{Rolling Covariance Estimate}
\usage{
rollCov(R, width)
}
\arguments{
\item{R}{xts or zoo object of asset returns}

\item{width}{width of rolling window}
}
\description{
This function calculates the covariance estimate between the returns of a
pair of assets over a rolling window.
}
\examples{
data(crsp_weekly)
R <- largecap_weekly[,1:4]
tail(rollCov(R, 10))
}
\author{
Ross Bennett
}
\seealso{
\code{\link{cov}}
}

