% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rollFUN.R
\name{rollSD}
\alias{rollSD}
\title{Rolling Standard Deviation Estimate}
\usage{
rollSD(R, width)
}
\arguments{
\item{R}{xts or zoo object of asset returns}

\item{width}{width of rolling window}
}
\description{
This function calculates the standard deviation estimate of asset returns
over a rolling window
}
\examples{
data(crsp_weekly)
R <- largecap_weekly[,1:4]
tail(rollSD(R, 10))
}
\author{
Ross Bennett
}
\seealso{
\code{\link{sd}}
}

