% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rollFUN.R
\name{rollSimpleVolatility}
\alias{rollSimpleVolatility}
\title{Rolling Simple Volatility Estimate}
\usage{
rollSimpleVolatility(R, width)
}
\arguments{
\item{R}{xts or zoo object of asset returns}

\item{width}{width of rolling window}
}
\description{
This function calculates the simple volatility estimate of asset returns
over a rolling window.
}
\details{
The simple volatility of x is defined as

\deqn{
  \sigma = \sqrt{\frac{1}{n} \sum_{i=1}^n x_i^2}
}
}
\examples{
data(crsp_weekly)
R <- largecap_weekly[,1:2]
tail(rollSimpleVolatility(R, 10))
}
\author{
Ross Bennett
}
\seealso{
\code{\link{sd}}, \code{\link{simpleVolatility}}
}

