\name{smallcap.ts}
\docType{data}
\alias{smallcap.ts}
\title{CRSP Small Cap Monthly Returns}
\description{
Monthly returns data of a selection of small cap stocks.
}
\usage{smallcap.ts}
\details{
CRSP Data used in GARPFRM and related publications with the kind permission of the Center for Research in Security Prices.

see \url{http://www.crsp.com/about-crsp}
}
\format{xts object with monthly observations}
\source{\url{http://www.crsp.com}}
\references{

}
\examples{
data(crsp.short)
head(smallcap.ts)
}
\keyword{datasets}


