% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/discountFactorArbitrage.R
\name{yieldCurveVasicek}
\alias{yieldCurveVasicek}
\title{Estimate Vasicek zero-coupon yield}
\usage{
yieldCurveVasicek(r, k, theta, sigma, maturity)
}
\arguments{
\item{r}{initial short rate}

\item{k}{speed of reversion parameter}

\item{theta}{long-term reversion yield}

\item{sigma}{randomness parameter. Modelled after Brownian Motion}

\item{maturity}{maturity of the bond}
}
\value{
yield curve estimate from Vasicek model
}
\description{
This function calculates the Vasicek yield given an initial data calibration
}
\author{
Thomas Fillebeen
}

