% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/riskMetricsAndHedges.R
\name{ytmSolve}
\alias{ytmSolve}
\title{Solve for the yield to maturity of a bond}
\usage{
ytmSolve(ytm, couponRate, m, nPayments, face, targetPrice)
}
\arguments{
\item{ytm}{yield to maturity}

\item{couponRate}{coupon rate}

\item{m}{compounding frequency}

\item{nPayments}{is the number of payments}

\item{face}{is the face value}

\item{targetPrice}{is the price of the bond}
}
\value{
Absolute value of difference between the price and the present value
}
\description{
This function solves for the yield to maturity of a fixed rate coupon bond
given the discount curve and bond data.
}
\author{
Thomas Fillebeen
}

