% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/discountFactorArbitrage.R
\name{bondSpec}
\alias{bondSpec}
\title{Constructor for bond specification}
\usage{
bondSpec(time = seq(from = 0.5, to = 2, by = 0.5), face = 100, m = 2,
  couponRate = 0.01)
}
\arguments{
\item{time}{vector of sequence of coupon payments in years}

\item{face}{face value of bond}

\item{m}{compounding frequency}

\item{couponRate}{rate the coupon pays}
}
\value{
a \code{bond} object with the bond data used for pricing
}
\description{
Create a bond specification.
}
\examples{
time = seq(from=0.5, to=2, by=0.5)
bond = bondSpec(time, face=100, m=2, couponRate = 0.0475)
}
\author{
Thomas Fillebeen
}

