% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/boot.R
\name{bootCov}
\alias{bootCov}
\title{Bootstrap Covariance}
\usage{
bootCov(R, ..., replications = 1000)
}
\arguments{
\item{R}{xts object or matrix of asset returns}

\item{replications}{number of bootstrap replications.}

\item{\dots}{passthrough parameters to \code{\link[stats]{cov}}}
}
\description{
Bootstrap the covariance of an xts object or matrix of asset returns
}
\examples{
data(crsp_weekly)
R <- largecap_weekly[,1:4]
bootCov(R[,1:2])
bootCov(R)
}
\author{
Ross Bennett
}
\seealso{
\code{\link[stats]{cov}}
}

