% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/boot.R
\name{bootSD}
\alias{bootSD}
\title{Bootstrap Standard Deviation}
\usage{
bootSD(R, ..., replications = 1000)
}
\arguments{
\item{R}{xts object or matrix of asset returns}

\item{replications}{number of bootstrap replications.}

\item{\dots}{passthrough parameters to \code{\link[stats]{sd}}}
}
\description{
Bootstrap the standard deviation of an xts object or matrix of asset returns
}
\examples{
data(crsp_weekly)
R <- largecap_weekly[,1:4]
bootSD(R[,1])
bootSD(R)
}
\author{
Ross Bennett
}
\seealso{
\code{\link[stats]{sd}}
}

