\name{chartSML}
\alias{chartSML}
\title{CAPM SML}
\usage{
  chartSML(object)
}
\arguments{
  \item{object}{a capm object created by
  \code{\link{CAPM}}}
}
\description{
  Description of CAPM Security Market Line (SML) SML is the
  represesentation of the CAPM. It illustrates the expected
  rate of return of an individual secuirty as a function of
  systematic, non-diversified risk (known as beta).
}

