% Generated by roxygen2 (4.0.1): do not edit by hand
\name{chartSML}
\alias{chartSML}
\title{CAPM SML}
\usage{
chartSML(object, ..., main = "Estimated SML")
}
\arguments{
\item{object}{a capm object created by \code{\link{CAPM}}.}

\item{\dots}{passthrough parameters to \code{\link{plot}}.}

\item{main}{a main title for the plot.}
}
\description{
Security Market Line (SML) of the CAPM.
The SML is a represesentation of the CAPM. It illustrates the expected rate
of return of an individual security as a function of systematic,
non-diversified risk (known as beta).
}
\author{
Thomas Fillebeen
}

