% Generated by roxygen2 (4.0.1): do not edit by hand
\name{computeGreeks}
\alias{computeGreeks}
\title{Option Greeks}
\usage{
computeGreeks(option, greek = "delta", prices = NULL, maturities = NULL,
  plot = FALSE, ...)
}
\arguments{
\item{option}{an \code{option} object created with \code{\link{optionSpec}}}

\item{greek}{one of "delta", "theta", "gamma", "rho", or "vega"}

\item{prices}{vector of values to compute the greeks as time
to maturity varies}

\item{maturities}{vector of values to compute the greeks as time
to maturity varies}

\item{plot}{TRUE/FALSE to plot the greek value as the underlying price and/ time to maturity vary}

\item{\dots}{passthrough parameters to \code{\link{plot}}}
}
\description{
Compute the greeks of an option using the Black-Scholes-Merton framework
}
\author{
Ross Bennett
}

