% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/options.R
\name{computeGreeks}
\alias{computeGreeks}
\alias{deltaBS}
\alias{deltaBS,}
\alias{gammaBS}
\alias{gammaBS,}
\alias{rhoBS}
\alias{thetaBS}
\alias{thetaBS,}
\alias{vegaBS}
\alias{vegaBS,}
\title{Option Greeks}
\usage{
computeGreeks(option, greek = c("delta", "theta", "gamma", "rho", "vega"),
  prices = NULL, maturities = NULL, plot = FALSE, ...)

deltaBS(S0, K, r, q, vol, ttm, type)

thetaBS(S0, K, r, q, vol, ttm, type)

gammaBS(S0, K, r, q, vol, ttm, type)

vegaBS(S0, K, r, q, vol, ttm, type)

rhoBS(S0, K, r, q, vol, ttm, type)
}
\arguments{
\item{option}{an \code{option} object created with \code{\link{optionSpec}}}

\item{greek}{one of "delta", "theta", "gamma", "rho", or "vega"}

\item{prices}{vector of values to compute the greeks as time
to maturity varies}

\item{maturities}{vector of values to compute the greeks as time
to maturity varies}

\item{plot}{TRUE/FALSE to plot the greek value as the underlying price and/ time to maturity vary}

\item{S0}{underlying asset price}

\item{K}{strike price}

\item{r}{risk free rate}

\item{q}{continuous dividend yield rate for options on stocks or stock
indices paying a dividend. Also the foreign risk free rate for options on
currencies}

\item{vol}{volatility of the underlying asset price}

\item{ttm}{tmie to maturity, the life of the option, measured in years}

\item{type}{type of the option; "call" or "put"}

\item{\dots}{passthrough parameters to \code{\link{plot}}}
}
\description{
Compute the greeks of an option using the Black-Scholes-Merton framework
}
\examples{
euro.call <- optionSpec(style="european", type="call", S0=30, K=30, maturity=1, r=0.05, volatility=0.25, q=0)
# European call greeks
computeGreeks(euro.call, greek = "delta")
computeGreeks(euro.call, greek = "gamma")
computeGreeks(euro.call, greek = "theta")
computeGreeks(euro.call, greek = "vega")
computeGreeks(euro.call, greek = "rho")

# Plotting
computeGreeks(euro.call, "delta", prices = seq(20, 40, 1), plot = TRUE)
computeGreeks(euro.call, "delta", maturities = seq(0.5, 0.01, -0.01), plot = TRUE)
}
\author{
Ross Bennett
}

