% Generated by roxygen2 (4.0.1): do not edit by hand
\name{discountFactor}
\alias{discountFactor}
\title{Estimate discountFactor}
\usage{
discountFactor(price, cashFlow)
}
\arguments{
\item{bond}{a \code{discountFactorArbitrage} object}

\item{price}{of a bond}
}
\value{
cashFlow of a bond
}
\description{
This function calculates the discountFactor (DF) given price
and cashFlows.
}
\author{
TF
}

