\name{endingPrices}
\alias{endingPrices}
\title{Ending Prices of Monte Carlo Simulation}
\usage{
  endingPrices(mc)
}
\arguments{
  \item{mc}{monte carlo object created with
  \link{\code{monteCarlo}}}
}
\value{
  vector ending prices
}
\description{
  Get the ending prices, i.e. terminal values, of a monte
  carlo simulation
}

