\name{EWMA}
\alias{EWMA}
\title{Exponential Weighted Moving Average (EWMA)}
\usage{
  EWMA(R, lambda = 0.94, inWnd = 10, cor = FALSE)
}
\arguments{
  \item{R}{}

  \item{lambda}{}

  \item{inWnd}{}

  \item{cor}{option (default = FALSE)}
}
\description{
  Description of EWMA. The function handles UV and MLM
  objects and returns either cov/cor.
}

