\name{EWMA}
\alias{EWMA}
\title{Exponential Weighted Moving Average (EWMA)}
\usage{
EWMA(R, lambda = 0.94, initialWindow = 10, cor = FALSE)
}
\arguments{
  \item{R}{}

  \item{lambda}{}

  \item{initialWindow}{is the initializing window}

  \item{correlation}{option (cor by default = FALSE)}
}
\description{
Description of EWMA. The function handles UV and MLM
objects and returns either cov/cor.
}

