% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/EWMA.R
\name{EWMA}
\alias{EWMA}
\title{EWMA Model}
\usage{
EWMA(R, lambda = 0.94, initialWindow = 10, n = 10,
  type = c("volatility", "covariance", "correlation"))
}
\arguments{
\item{R}{xts object of asset returns}

\item{lambda}{smoothing parameter, must be greater than 0 or less than 1. If
NULL, lambda will be estimated by minimizing the mean squared error
between the estimated value and the realized value.}

\item{initialWindow}{initial window of observations used in estimating the
initial conditions}

\item{n}{number of periods used to calculate realized volatility, covariance, or correlation.}

\item{type}{estimate volatility, covariance, or correlation.}
}
\value{
an EWMA object with the following elements
\itemize{
  \item \code{estimate} EWMA model estimated statistic
  \item \code{model} list with model parameters
  \item \code{data} list with original returns data and realized statistic if applicable
}
}
\description{
EWMA model to estimate volatility, covariance, and correlation
}
\details{
If lambda=NULL, the lambda value can be estimated for univariate estimates
of volatility,  covariance, and correlation by minimizing the mean
squared error between the estimated value and realized value.
}
\examples{
# data and parameters for EWMA estimate
data(crsp_weekly)
R <- largecap_weekly[, 1:2]
mvR <- largecap_weekly[,1:4]
lambda <- 0.94
initialWindow <- 150

# volatility estimate of univariate data
lambda <- estimateLambdaVol(R[,1], initialWindow, n=10)
vol1 <- EWMA(R[,1], lambda=NULL, initialWindow, n=10, "volatility")
vol1a <- EWMA(R[,1], lambda, initialWindow, n=10, "volatility")
all.equal(vol1$estimate, vol1a$estimate)
vol1

# covariance estimate of bivariate data
lambda <- estimateLambdaCov(R, initialWindow, n=10)
cov1 <- EWMA(R, lambda=NULL, initialWindow, n=10, "covariance")
cov1a <- EWMA(R, lambda, initialWindow, n=10, "covariance")
all.equal(cov1$estimate, cov1a$estimate)
cov1

# correlation estimate of bivariate data
lambda <- estimateLambdaCor(R, initialWindow, n=10)
cor1 <- EWMA(R, lambda=NULL, initialWindow, n=10, "correlation")
cor1a <- EWMA(R, lambda, initialWindow, n=10, "correlation")
all.equal(cor1$estimate, cor1a$estimate)
cor1

# Multivariate EWMA estimate of covariance
lambda <- 0.94
cov_mv <- EWMA(mvR, lambda, initialWindow, type="covariance")
cov_mv

# Multivariate EWMA estimate of correlation
cor_mv <- EWMA(mvR, lambda, initialWindow, type="correlation")
cor_mv
}
\author{
Ross Bennett and Thomas Fillebeen
}

