% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generic_forecast.R
\name{forecast.uvEWMAvol}
\alias{forecast.uvEWMAvol}
\title{Forecast Univariate EWMA Volatility Model}
\usage{
\method{forecast}{uvEWMAvol}(model, nAhead = 1, ...)
}
\arguments{
\item{model}{EWMA model fit via \code{\link{EWMA}}}

\item{nAhead}{number of steps ahead to forecast. (nAhead = 1 only supported)}

\item{\dots}{additional passthrough parameters}
}
\value{
one period ahead EWMA volatility forecast
}
\description{
Forecasting for EWMA Volatility models fit via \code{\link{EWMA}}
}

