\name{GARPFRM-package}
\alias{GARPFRM-package}
\alias{GARPFRM}
\docType{package}
\title{
Functions to implement the topics presented in 'Financial Risk Manager (FRM) Part 1' (2012) series of books.
}
\description{
TODO: Better description
This package provides a framework to implement the topics and ideas presented in 'Financial Risk Manager (FRM) Part 1: Quantitative Analysis' (2012) and 'Financial Risk Manager (FRM) Part 1: Foundations of Risk Management' (2012)
}

\details{
\tabular{ll}{
Package: \tab GARPFRM\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2013-11-17\cr
License: \tab What license is it under?\cr
}

An overview of how to use the package, including the most important functions.
Discuss each section, borrowing text from vignettes. Also make sure that each area of functions is covered.
\itemize{
  \item Delineating Efficient Frontier
  \item CAPM
  \item Performance Measures
  \item Monte Carlo
  \item Bootstrap
  \item EWMA
  \item GARCH
  \item VaR
}

}
\author{
Ross Bennett and Thomas Fillebeen

Maintainer: Thomas Fillebeen <tdf17@uw.edu>
}
\references{
TODO: Add references
rugarch
PerformanceAnalytics
xts
GARP books
}

\keyword{ package }
\seealso{
% Optional links to other man pages, e.g.
% \code{\link[<pkg>:<pkg>-package]{<pkg>}}
}
\examples{
% simple examples of the most important functions
}
