\name{getCor}
\alias{getCor}
\title{EWMA Correlation}
\usage{
  getCor(EWMA, assets)
}
\arguments{
  \item{object}{an EWMA object created by \code{EWMA}}

  \item{assets}{character vector or numeric vector. The
  assets can be specified by name or index.}
}
\description{
  Extract the correlation of two assets from an \code{EWMA}
  object
}

