% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/EWMA.R
\name{getCov}
\alias{getCov}
\title{EWMA Covariance}
\usage{
getCov(EWMA, assets)
}
\arguments{
\item{EWMA}{an EWMA object created by \code{\link{EWMA}}}

\item{assets}{character vector or numeric vector. The assets can be
specified by name or index.}
}
\description{
Extract the covariance of two assets from an \code{mvEWMAcov} object
}
\examples{
data(crsp_weekly)
mvR <- largecap_weekly[,1:4]
lambda <- 0.94
initialWindow <- 150
cov_mv <- EWMA(mvR, lambda, initialWindow, type="covariance")
# Extract the estimated covariance between ORCL and MSFT
tail(getCov(cov_mv, assets=c("ORCL", "MSFT")))
}
\author{
Ross Bennett and Thomas Fillebeen
}

