\name{getEstimate}
\alias{getEstimate}
\title{Get Estimated Values}
\usage{
getEstimate(model, ...)
}
\arguments{
  \item{object}{fitted model (currently only EWMA)}

  \item{/dots}{passthrough parameters (not currently used)}
}
\value{
model estimate
}
\description{
Get the estimated values from the model
}
\author{
Ross Bennett
}

