% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getFit}
\alias{getFit}
\title{Get Fitted GARCH Model}
\usage{
getFit(garch)
}
\arguments{
\item{garch}{a GARCH model specification and fit created with \code{uvGARCH}}
}
\value{
an object of class uGARCHfit
}
\description{
Function to extract the fitted GARCH model object
}

