% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/garch11.R
\name{getSpec}
\alias{getSpec}
\title{Get GARCH Model Specification}
\usage{
getSpec(garch)
}
\arguments{
\item{garch}{a GARCH model specification and fit created with \code{uvGARCH}}
}
\value{
an object of class uGARCHspec
}
\description{
Function to extract the GARCH model specification object
}

