% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/backTestVaR.R
\name{getVaREstimates}
\alias{getVaREstimates}
\title{VaR Estimates
Extract VaR Estimates from a VaR Backtest}
\usage{
getVaREstimates(object, ...)
}
\arguments{
\item{object}{an object created by \code{\link{backtestVaR}}.}

\item{\dots}{not currently used}
}
\value{
xts object of unconditional VaR estimates
}
\description{
VaR Estimates
Extract VaR Estimates from a VaR Backtest
}
\author{
Ross Bennett
}
\seealso{
\code{\link{backtestVaR}}
}

