\name{hypTest}
\alias{hypTest}
\title{CAPM Hypothesis Test}
\usage{
  hypTest(object, CI)
}
\arguments{
  \item{object}{a capm object created by
  \code{\link{CAPM}}}

  \item{CI}{confidence level}
}
\description{
  Description of CAPM beta/alpha hypothesis test TODO: We
  need to clearly define the null hypothesis here
}
\details{
  Generalization is termed a two-sided or two-tailed test.
  Returns a true (reject) or false (fail to reject).
}

