\name{hypTest}
\alias{hypTest}
\title{CAPM hypTest}
\usage{
hypTest(object, CI)
}
\arguments{
  \item{object}{a capm object created by
  \code{\link{CAPM}}}
}
\description{
Description of CAPM beta/alpha hypothesis test
Generalization is termed a two-sided or two-tailed test.
Returns a true (reject) or false (fail to reject).
}

