% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/discountFactorArbitrage.R
\name{is.bond}
\alias{is.bond}
\title{To determine if user is specifying bond parameters correctly}
\usage{
is.bond(object)
}
\arguments{
\item{object}{a bond specification object created by \code{\link{bondSpec}}}
}
\description{
To determine if user is specifying bond parameters correctly
}
\author{
Thomas Fillebeen
}

