% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/options.R
\name{optionValue}
\alias{optionValue}
\title{Option Value}
\usage{
optionValue(option, method = c("Binomial", "Black-Scholes"), N = 20,
  verbose = FALSE, ...)
}
\arguments{
\item{option}{an \code{option} object created with \code{\link{optionSpec}}}

\item{method}{the method used to value the option}

\item{N}{number of steps in binomial tree}

\item{verbose}{TRUE/FALSE default FALSE. TRUE prints the node information
of the binomial tree}

\item{\dots}{any other passthrough parameters}
}
\value{
the estimated value of the option
}
\description{
Estimate the value of an option
}
\examples{
am.call <- optionSpec(style="american", type="call")
am.call.val <- optionValue(am.call, N=4)
euro.call <- optionSpec(style="european", type="call", S0=30, K=30, maturity=1, r=0.05, volatility=0.25, q=0)
euro.call.val.bs <- optionValue(euro.call, method="Black-Scholes")
}
\author{
Ross Bennett
}

