% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/capm.R
\name{plot.capm_mlm}
\alias{plot.capm_mlm}
\title{Plotting method for CAPM}
\usage{
\method{plot}{capm_mlm}(x, y, ..., main = "CAPM")
}
\arguments{
\item{x}{a capm object created by \code{\link{CAPM}}.}

\item{y}{not used}

\item{main}{a main title for the plot}

\item{\dots}{passthrough parameters to \code{\link{plot}}.}
}
\description{
Plot a fitted CAPM object
}
\author{
Thomas Fillebeen
}

