\name{plot.capm_uv}
\alias{plot.capm_uv}
\title{Plotting method for CAPM}
\usage{
\method{plot}{capm_uv}(x, y, number, ..., main = "CAPM")

\method{plot}{pca loadings}(x, y, number, ..., main = "CAPM")
}
\arguments{
  \item{x}{a capm object created by \code{\link{CAPM}}.}

  \item{y}{not used}

  \item{\dots}{passthrough parameters to
  \code{\link{plot}}.}

  \item{main}{a main title for the plot}

  \item{x}{a PCA object created.}

  \item{y}{not used}

  \item{number}{specify the nunber of loadings}

  \item{\dots}{passthrough parameters to
  \code{\link{plot}}.}

  \item{main}{a main title for the plot}
}
\description{
Plot a fitted CAPM object

Plot a fitted PCA object
}
\author{
Thomas Fillebeen

Thomas Fillebeen
}

