% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.PCA}
\alias{plot.PCA}
\title{Plotting method for PCA}
\usage{
\method{plot}{PCA}(x, y, ..., main = "Beta from PCA regression")
}
\arguments{
\item{x}{a PCA object created by \code{\link{PCA}}}

\item{y}{not used}

\item{number}{specify the nunber of loadings}

\item{\dots}{passthrough parameters to \code{\link{plot}}.}

\item{main}{a main title for the plot}
}
\description{
Plot a fitted PCA object
}
\author{
Thomas Fillebeen
}

