% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/riskMetricsAndHedges.R
\name{plot.PCA}
\alias{plot.PCA}
\title{Plotting method for PCA}
\usage{
\method{plot}{PCA}(x, y, ..., main = "Beta from PCA regression",
  separate = TRUE)
}
\arguments{
\item{x}{a PCA object created by \code{\link{PCA}}}

\item{y}{not used}

\item{main}{a main title for the plot}

\item{separate}{if TRUE plot of same, and if FALSE plot separately}

\item{\dots}{passthrough parameters to \code{\link{plot}}.}
}
\description{
Plot a fitted PCA object
}
\author{
Thomas Fillebeen
}

