% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.uvGARCH}
\alias{plot.uvGARCH}
\title{Plot GARCH Model}
\usage{
\method{plot}{uvGARCH}(x, y, ..., which)
}
\arguments{
\item{x}{uvGARCH object create via \code{uvGARCH}}

\item{y}{not used}

\item{\dots}{additional parameters passed to plot method for uGARCHfit objects}

\item{which}{plot selection}
}
\description{
Plots for fitted GARCH Models
}

