% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rollFUN.R
\name{rollCor}
\alias{rollCor}
\title{Rolling Correlation Estimate}
\usage{
rollCor(R, width)
}
\arguments{
\item{R}{xts or zoo object of asset returns}

\item{width}{width of rolling window}
}
\description{
This function calculates the correlation estimate between the returns of a
pair of assets over a rolling window.
}
\examples{
data(crsp_weekly)
R <- largecap_weekly[,1:4]
tail(rollCor(R, 10))
}
\author{
Ross Bennett
}
\seealso{
\code{\link{cor}}
}

